+ GameClient added + tours added + game result added + other minor changes and bugfix

+++ Jwt.h
@@ -2,16 +2,17 @@
 #define _JWT_H_
 
 #include <string>
+#include <stdint.h>
 
 #define TOKEN_LIFETIME_SEC 1800 // 30 min
 
 class Jwt {
     public:
     Jwt(std::string token, std::string key);
-    Jwt(long user_id, long expire, std::string key);
+    Jwt(uint64_t user_id, long expire, std::string key);
     std::string get_token();
     bool verify();
-    int get_uid();
+    uint64_t get_uid();
     
     private:
     std::string _header;
